1.0.0

+++ main/models.py
@@ -3,7 +3,7 @@ kinds = [('l','lion'),('e','elephant')]
 # Create your models here.
 class Positions(models.Model):
     name = models.CharField(max_length=40)
-    wage = models.IntegerField()
+    wage = models.FloatField()
 
     def __str__(self):
         return self.name
